Move basemap function to basemap.py
