+++ pygmt/src/__init__.py
@@ -2,14 +2,12 @@
 Source code for PyGMT modules.
 """
 # pylint: disable=import-outside-toplevel
-# GMT Primary modules
+from pygmt.src.basemap import basemap
 from pygmt.src.coast import coast
 from pygmt.src.colorbar import colorbar
 from pygmt.src.grdcontour import grdcontour
 from pygmt.src.image import image
 from pygmt.src.legend import legend
 from pygmt.src.logo import logo
-
-# GMT Supplementary modules
 from pygmt.src.meca import meca
 from pygmt.src.text import text
